+++ CHANGELOG.md
@@ -19,6 +19,7 @@
 - chore: move benchmarks to tools folder
 - chore: simplify mqttPacket encoder/decoder exports
 - fix(server): fix empty payload to clear retained message
+- fix(client): fix client handling of broker V5 rejects
 - fix: remove password from demo logging and debug logging
 
 ## [v1.11.3] 31-05-2026